Migrated the br2 integration macro
